lantern/django/django_intro/requirements/base.txt
asgiref==3.2.7
Django==3.0.5
pytz==2020.1
sqlparse==0.3.1
django-environ==0.4.5
psycopg2-binary==2.8.4
Pillow==7.1.2
djangorestframework==3.11.0
